Updated Crawlers

+++ Crawler/util/sqlite_connector.py
@@ -4,21 +4,42 @@ import sys
 connection = None
 table_name = None
 tick = 0
+table_vartype = ""
 
-def open_database_connection(table_schema, database, table):
-    global connection
-    global table_name
-    table_name = table
-    connection = lite.connect('database/%s.db' % database)
+def open_database_connection(create_database, table_schema, database, table, vartype):
+    if create_database:
+        global connection
+        global table_name
+        global table_vartype
+        table_vartype = vartype.split(", ")
+        table_name = table
+        connection = lite.connect('database/%s.db' % database)
 
-    cur = connection.cursor()
-    cur.execute("DROP TABLE IF EXISTS temp_%s;" % table_name)
-    connection.commit()
-    cur.execute("DROP TABLE IF EXISTS temp_%s" % table_name)
-    cur.execute("CREATE TABLE temp_%s(%s)" % (table_name, table_schema))
-    cur.execute("CREATE TABLE IF NOT EXISTS %s(%s)" % (table_name, table_schema))
-    cur.close()
-    connection.commit()
+        cur = connection.cursor()
+        cur.execute("DROP TABLE IF EXISTS temp_%s;" % table_name)
+        connection.commit()
+        cur.execute("DROP TABLE IF EXISTS temp_%s" % table_name)
+        cur.execute("CREATE TABLE temp_%s(%s)" % (table_name, table_schema))
+        cur.execute("CREATE TABLE IF NOT EXISTS %s(%s)" % (table_name, table_schema))
+        cur.close()
+        connection.commit()
+    else:
+        connection = lite.connect('database/%s.db' % database)
+        cur = connection.cursor()
+        cur.execute("SELECT %s FROM %s;" % (table_schema, database))
+        data = cur.fetchall()
+        variable_names = table_schema.split(", ")
+        data_list = []
+        print("Fetching All data..")
+        for row in data:
+            tick = 0
+            section = {}
+            for var in variable_names:
+                section[var] = row[tick]
+                tick += 1
+            data_list.append(section)
+        print("fetched all data!")
+        return data_list
     return True
 
 def write_to_database(data, database_layout):
@@ -27,15 +48,15 @@ def write_to_database(data, database_layout):
     data_values = ""
 
     # Gets the name of each variable in the dictionary from the layout and concatenates the values together.
+    tick = 0
     for data_name in names:
-        data_type = "\"%s\""
-        if str(data[data_name]).isdigit() and data_name != "age":
-            data_type = "\"%d\""
+        data_type = "%s" % table_vartype[tick]
 
         if data_values == "":
-            data_values += data_type % data[data_name]
+            data_values +=  ("\"" + data_type + "\"") % data[data_name]
         else:
-            data_values += ", " + data_type % data[data_name]
+            data_values += (", \"" + data_type + "\"") % data[data_name]
+        tick += 1
     
     global tick;
     tick += 1;
